Rename com1DFA/com1DFAPy for switch
