+++ avaframe/com1DFA/DFAfunctionsCython.pyx
