+++ avaframe/com1DFA/com1DFACfg.ini
@@ -4,46 +4,171 @@
 
 
 [GENERAL]
-# Path to com1DFA executable and AK_Attributes files
-com1Exe = ./com1DFA.exe -files files/AK_Attributes/
-# True if full output shall be printed to terminal
-fullOut = False
+# acceleration of gravity [m/s²]
+gravAcc = 9.81
+# density of snow [kg/m³]
+rho = 200
+rhoEnt = 100
+# mass per particle [kg]
+massPerPart = 1250
+# release thickness per particle
+deltaTh = 0.25
+# minimum flow depth [m]
+hmin = 0.05
+# curvature acceleration coefficient
+# take curvature term into account in the gravity acceleration term
+# 0 if deactivated, 1 if activated
+curvAcceleration = 1
+
+#+++++++++++++SPH parameters
+# SPH gradient option
+# 0) No pressure gradients computed
+# 1) SamosAT style (no reprojecion on the surface, dz = 0 and gz is used)
+# 2) SamostAt but done in the local coord system (will hopefully allow us to add the earth pressure coef)
+# and this time reprojecion on the surface, dz not 0 and g3 is used
+# 3) same as 2) but correcting the gradient to get constant exact gradients
+# 4) same as 2) but correcting the gradient to get linear exact gradients
+sphOption = 1
+# minimum SPH distance [m]
+minRKern = 0.001
+# sph kernel smoothing length [m]
+sphKernelRadius = 5
+# number of particles defined by: MPPDIR= mass per particle direct, MPPDH= mass per particles through release thickness
+massPerParticleDeterminationMethod = MPPDH
+
+velMagMin = 1.0e-6
+depMin = 1.0e-6
+
+
+#+++++++++++++Mesh and interpolation
+# interpolation option
+# 3 Options available : -0: nearest neighbour interpolation
+#                       -1: equal weights interpolation
+#                       -2: bilinear interpolation
+interpOption = 2
+
+# remesh the input DEM
+# expected mesh size [m]
+meshCellSize = 5
+# threshold under which no remeshing is done
+meshCellSizeThreshold = 0.001
+
+# Normal computation on rectangular grid
+# 4 triangles method        6 triangles method         8 triangles method
+# +----U----UR---+---+--... +----+----+----+---+--... +----+----+----+---+--...
+# |   /|\   |   /|          |   /| 2 /|   /|          |\ 2 | 3 /|   /|           Y
+# |  / | \  |  / |          |  / |  / |  / |          | \  |  / |  / |           ^
+# | /  |  \ | /  | /        | /  | /  | /  | /        |  \ | /  | /  | /         |
+# |/ 1 | 2 \|/   |/         |/ 1 |/ 3 |/   |/         | 1 \|/ 4 |/   |/          |
+# +----P----L----+---+--... +----*----+----+---+--... +----*----+----+----+--... +-----> X
+# |\ 4 | 3 /|   /|          | 6 /| 4 /|   /|          | 8 /|\ 5 |   /|
+# | \  |  / |  / |          |  / |  / |  / |          |  / | \  |  / |
+# |  \ | /  | /  | /        | /  | /  | /  | /        | /  |  \ | /  | /
+# |   \|/   |/   |/         |/ 5 |/   |/   |/         |/ 7 | 6 \|/   |/
+# +----+----+----+---+--... +----+----+----+---+--... +----+----+----+---+--...
+# |   /|   /|   /|          |   /|   /|   /|          |   /|   /|   /|
+# 4 Options available : -1: simple cross product method (with the diagonals P-UR and U-L)
+#                       -4: 4 triangles method
+#                       -6: 6 triangles method
+#                       -8: 8 triangles method
+methodMeshNormal = 1
+
+# subgridMixingFactor
+subgridMixingFactor = 100.
+
+# release thickness
+relTh = 1.
+# Add secondary release area
+secRelArea = False
+# secondary area release thickness
+secondaryRelTh = 0.5
+
+#++++++++++++Time stepping parameters
+# fixed time step [s]
+dt = 0.1
+# End time [s]
+tEnd = 400
+# to use a variable time step (defined by the CFL number)
+cflTimeStepping = False
+# courant number
+cMax = 0.5
+# max time step in case of small velocities
+maxdT = 0.5
+# to constrain the lower margin of the allowed time step
+constrainCFL = False
+# min time step in case of high velocity
+mindT = 0.01
+
+# stopCriterion (0.01, stops when ke<0.01*pke)
+stopCrit = 0.01
+
+
+# friction type (samosAT, Coulomb)
+frictModel = samosAT
+#+++++++++++++SamosAt friction model
+mu = 0.15500
+Rs0 = 0.222
+kappa = 0.43
+R = 0.05
+B = 4.13
+
+#++++++++++++ Resistance force parameters
+# height of the obstacles
+hRes = 10
+# resistance coeff
+cw = 0.5
+# diameter of obstacles
+dRes = 0.3
+# spacing between obstacles
+sres = 5
+#++++++++++++ entrainment Erosion Energy
+hEnt = 0.3
+entEroEnergy = 5000
+entShearResistance = 0
+entDefResistance = 0
+
+
+#+++++++++++++ Output
+# desired result Parameters (ppr, pfd, pfv, FD, FV, P, particles)
+resType = ppr_pfd_pfv
+# saving time step, i.e.  time in seconds (first and last time step are always saved)
+# option 1: give an interval with start:interval in seconds (tStep = 0:5 - this will save desired results every 5 seconds for the full simulation)
+# option 2: explicitly list all desired time steps (closest to actual computational time step) separated by | (example tSteps = 1|50.2|100)
+# NOTE: initial and last time step are always saved!
+tSteps = 1
+
+
+#++++++++++++++++ particle Initialisation
+# if true use file (import particles initial distribution from file)
+initialiseParticlesFromFile = False
+particleFile =
+# seed for random generator
+seed = 12345
 
 #++++++++++++++++ simulation type
 # list of simulations that shall be performed (null, ent, res, entres, available (use all available input data))
 simTypeList = available
+# model type
+modelType = dfa
 
-#++++++++++++++++ release area and time step output
+[VISUALISATION]
+# if particle properties shall be exported to csv files
+writePartToCSV = False
+# particle properties to be saved to csv (options: ux, uy, uz, velocityMagnitude,..)
+particleProperties = velocityMagnitude|m
+
+
+[FLAGS]
+# True if dev release areas
+flagDev = False
 # specify a particular release area scenario, provide name of shapefile with or without extension .shp (optional)
 releaseScenario =
-# if True: flow depth of first and last time step are exported
-addTSteps = True
 
 
-[PARAMETERVAR]
-# True if parameter variation option shall be included
-parameterVar = False
-# True if entres simulation shall be used for parameter variation
-varEnt = True
-# parameter to be varied, current option: Mu (friction parameter), RelTh (release area thickness)
-varPar = Mu
-# list of parameter values, individual values need to be separated by '|' or provide start:end:numberOfSteps
-varParValues = 0.05500|0.15500
-# Name of run cint file template if parameter variation is used
-varRunCint = vary
-
-
-[ENTRAINMENT]
-# True if entrainment depth shall be varied from default
-setEntDepth = False
-# set entrainment thickness
-entH =
-
-
-[DEFVALUES]
-# standard value for Mu
-Mu = 0.15500
-# standard value for release area thickness
-RelTh = 1.00000
-# standard value for entrainment thickness
-defaultEntH = 0.3
+[REPORT]
+# which result parameters shall be included as plots in report
+plotFields = ppr_pfd_pfv
+# units for output variables
+unitppr = kPa
+unitpfd = m
+unitpfv = ms-1
