+++ avaframe/com1DFA/simiSol.py
