+++ avaframe/com1DFA/timeDiscretizations.py
