+++ avaframe/com1DFAOrig/__init__.py
@@ -1,0 +1,1 @@
+"""Dense flow avalanche kernel (wrapper around c++ code)"""
