+++ avaframe/com1DFAOrig/com1DFACfg.ini
@@ -1,0 +1,49 @@
+### Config File - This file contains the main settings for the simulation run
+## Copy to  local_com1DFACfg.ini and set you parameters
+# This file is part of Avaframe.
+
+
+[GENERAL]
+# Path to com1DFA executable and AK_Attributes files
+com1Exe = ./com1DFA.exe -files files/AK_Attributes/
+# True if full output shall be printed to terminal
+fullOut = False
+
+#++++++++++++++++ simulation type
+# list of simulations that shall be performed (null, ent, res, entres, available (use all available input data))
+simTypeList = available
+
+#++++++++++++++++ release area and time step output
+# specify a particular release area scenario, provide name of shapefile with or without extension .shp (optional)
+releaseScenario =
+# if True: flow depth of first and last time step are exported
+addTSteps = True
+
+
+[PARAMETERVAR]
+# True if parameter variation option shall be included
+parameterVar = False
+# True if entres simulation shall be used for parameter variation
+varEnt = True
+# parameter to be varied, current option: Mu (friction parameter), RelTh (release area thickness)
+varPar = Mu
+# list of parameter values, individual values need to be separated by '|' or provide start:end:numberOfSteps
+varParValues = 0.05500|0.15500
+# Name of run cint file template if parameter variation is used
+varRunCint = vary
+
+
+[ENTRAINMENT]
+# True if entrainment depth shall be varied from default
+setEntDepth = False
+# set entrainment thickness
+entH =
+
+
+[DEFVALUES]
+# standard value for Mu
+Mu = 0.15500
+# standard value for release area thickness
+RelTh = 1.00000
+# standard value for entrainment thickness
+defaultEntH = 0.3
