--- avaframe/com1DFAPy/__init__.py
@@ -1,1 +1,0 @@
-"""Development code for python version of com1DFA"""
